Allow multiseries requests
Newish to python and git and api requests, so my solution may not be advisable.  I was unable to make a multiseries request, but specifying that the post request was json format fixes the issue.  Code that did not work prior to PR: 

```python
# registration key will need to be specified
jdat = bls.get_series(["SUUR0000SA0", "CUUR0000SA0"], startyear=2011, endyear=2011, key=regkey)
```